requirements.txt
dotenv
fastapi
uvicorn
boto3
requests
google-generativeai
google-cloud-aiplatform[adk,agent-engines]
google-adk
